revert the previous commit
Implementing nfree() as a function as it was made,
was erroneous. Setting the pointer to NULL did not
work because it could not be passed back to the
caller. On the other hand, this setting the pointer
to NULL is a kludge, it just masks bugs in the code.
Since there are 1211 entries of nfree(), there may
be too many bugs to fix. It is easier for me just
to revert the commit.